pyproject.toml
[tool.poetry]
name = "vnp-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
python-telegram-bot = "^21.4"
celery = "^5.4.0"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
django-celery-beat = "^2.6.0"
psycopg2 = "^2.9.9"
pydantic = "^2.9.1"
aiohttp = "^3.10.5"
aiofiles = "^24.1.0"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.4.6"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
